refactor(routing): consolidate JSON-RPC error response builders
- Created internal/routing/response_errors.go.
- Moved tool-level error builders from mcp_request.go and protocol-level
  error builders from guardrails.go into the new file.
- Renamed functions to explicitly distinguish between ToolExecutionErrors
  and ProtocolRejections to prevent confusion.
- Updated all references across the routing package.
Resolves #1467
