+++ internal/routing/mcp_request.go
@@ -267,50 +267,7 @@ type ElicitationInfo struct {
 	ElicitationID string
 }
 
-// SseJSONRPC constructs sse json-rpc event with custom body
-func SseJSONRPC(requestID any, writeBody func(b *strings.Builder)) string {
-	var b strings.Builder
-	b.WriteString("\nevent: message\ndata: {\"jsonrpc\":\"2.0\",\"id\":")
-	idBytes, err := json.Marshal(requestID)
-	if err != nil {
-		b.WriteString("null")
-	} else {
-		b.Write(idBytes)
-	}
-	writeBody(&b)
-	b.WriteString("\n\n")
-	return b.String()
-}
 
-// BuildSSEToolError constructs sse error response for tool call
-func BuildSSEToolError(requestID any, message string) string {
-	return SseJSONRPC(requestID, func(b *strings.Builder) {
-		b.WriteString(",\"result\":{\"content\":[{\"type\":\"text\",\"text\":")
-		b.WriteString(jsonQuote(message))
-		b.WriteString("}],\"isError\":true}}")
-	})
-}
-
-// BuildJSONToolError constructs a plain JSON-RPC error response for 2026-07-28
-func BuildJSONToolError(requestID any, message string) string {
-	var b strings.Builder
-	b.WriteString("{\"jsonrpc\":\"2.0\",\"id\":")
-	idBytes, err := json.Marshal(requestID)
-	if err != nil {
-		b.WriteString("null")
-	} else {
-		b.Write(idBytes)
-	}
-	b.WriteString(",\"result\":{\"content\":[{\"type\":\"text\",\"text\":")
-	b.WriteString(jsonQuote(message))
-	b.WriteString("}],\"isError\":true}}")
-	return b.String()
-}
-
-func jsonQuote(s string) string {
-	b, _ := json.Marshal(s)
-	return string(b)
-}
 
 // ResourceAuthority extracts the authority segment (host) from a resource URI.
 // For malformed URIs, returns the URI unchanged.
