+++ internal/routing/router_202511.go
@@ -131,7 +131,7 @@ func (r *Router202511) routeToolCall(ctx context.Context, table RoutingTable, mc
 		return &Decision{
 			Error: &Error{
 				StatusCode: 200,
-				JSONRPCErr: BuildSSEToolError(mcpReq.ID, "MCP error -32602: Tool not found"),
+				JSONRPCErr: BuildSSEToolExecutionError(mcpReq.ID, "MCP error -32602: Tool not found"),
 			},
 			SetHeaders: map[string]string{
 				SessionHeader: mcpReq.GetSessionID(),
@@ -146,7 +146,7 @@ func (r *Router202511) routeToolCall(ctx context.Context, table RoutingTable, mc
 		return &Decision{
 			Error: &Error{
 				StatusCode: 200,
-				JSONRPCErr: BuildSSEToolError(mcpReq.ID, "MCP error -32602: Tool not found"),
+				JSONRPCErr: BuildSSEToolExecutionError(mcpReq.ID, "MCP error -32602: Tool not found"),
 			},
 			SetHeaders: map[string]string{
 				SessionHeader: mcpReq.GetSessionID(),
@@ -199,7 +199,7 @@ func (r *Router202511) routeToolCall(ctx context.Context, table RoutingTable, mc
 				return &Decision{
 					Error: &Error{
 						StatusCode: 200,
-						JSONRPCErr: BuildSSEToolError(mcpReq.ID, tokenErr.Error()),
+						JSONRPCErr: BuildSSEToolExecutionError(mcpReq.ID, tokenErr.Error()),
 					},
 					SetHeaders: map[string]string{
 						SessionHeader: mcpReq.GetSessionID(),
@@ -259,7 +259,7 @@ func (r *Router202511) routePromptGet(ctx context.Context, table RoutingTable, m
 		return &Decision{
 			Error: &Error{
 				StatusCode: 200,
-				JSONRPCErr: "\nevent: message\ndata: {\"error\":{\"code\":-32602,\"message\":\"Prompt not found\"},\"jsonrpc\":\"2.0\"}\n\n",
+				JSONRPCErr: BuildSSEProtocolRejection(mcpReq.ID, -32602, "Prompt not found"),
 			},
 			SetHeaders: map[string]string{
 				SessionHeader: mcpReq.GetSessionID(),
@@ -274,7 +274,7 @@ func (r *Router202511) routePromptGet(ctx context.Context, table RoutingTable, m
 		return &Decision{
 			Error: &Error{
 				StatusCode: 200,
-				JSONRPCErr: BuildSSEToolError(mcpReq.ID, "MCP error -32602: Prompt not found"),
+				JSONRPCErr: BuildSSEToolExecutionError(mcpReq.ID, "MCP error -32602: Prompt not found"),
 			},
 			SetHeaders: map[string]string{
 				SessionHeader: mcpReq.GetSessionID(),
@@ -340,7 +340,7 @@ func (r *Router202511) routeResourceRead(ctx context.Context, table RoutingTable
 		return &Decision{
 			Error: &Error{
 				StatusCode: 200,
-				JSONRPCErr: BuildSSEToolError(mcpReq.ID, "MCP error -32602: Resource not found"),
+				JSONRPCErr: BuildSSEToolExecutionError(mcpReq.ID, "MCP error -32602: Resource not found"),
 			},
 			SetHeaders: map[string]string{
 				SessionHeader: mcpReq.GetSessionID(),
@@ -355,7 +355,7 @@ func (r *Router202511) routeResourceRead(ctx context.Context, table RoutingTable
 		return &Decision{
 			Error: &Error{
 				StatusCode: 200,
-				JSONRPCErr: BuildSSEToolError(mcpReq.ID, "MCP error -32602: Resource not found"),
+				JSONRPCErr: BuildSSEToolExecutionError(mcpReq.ID, "MCP error -32602: Resource not found"),
 			},
 			SetHeaders: map[string]string{
 				SessionHeader: mcpReq.GetSessionID(),
